+++ subsys/nrf_security/src/drivers/cracen/cracenpsa/src/hash.c
@@ -12,9 +12,7 @@
 #include <string.h>
 #include <sxsymcrypt/hash.h>
 #include <sxsymcrypt/internal.h>
-#include <sxsymcrypt/sha1.h>
-#include <sxsymcrypt/sha2.h>
-#include <sxsymcrypt/sha3.h>
+#include <sxsymcrypt/hashdefs.h>
 #include <cracen/statuscodes.h>
 #include <zephyr/sys/__assert.h>
 #include "common.h"
