+++ subsys/nrf_security/src/drivers/cracen/sicrypto/src/ed25519.c
@@ -33,7 +33,7 @@
  */
 
 #include <string.h>
-#include <sxsymcrypt/sha2.h>
+#include <sxsymcrypt/hashdefs.h>
 #include <silexpk/core.h>
 #include <silexpk/ed25519.h>
 #include <sicrypto/sicrypto.h>
